+++ packages/ui/src/components/ChainIcon/index.tsx
@@ -42,6 +42,7 @@ import kyveIconBoth from 'shared-utils/assets/icons/kyve-both.svg?url';
 import humansaiIconDark from 'shared-utils/assets/icons/humansai-both.svg?url';
 import gitopiaIconDark from 'shared-utils/assets/icons/gitopia-both.svg?url';
 import neutronIconBoth from 'shared-utils/assets/icons/neutron-both.svg?url';
+import jackalIconBoth from 'shared-utils/assets/icons/jackal-both.svg?url';
 import agoricLogoLight from 'shared-utils/assets/logos/agoric-light.png';
 import akashLogoDark from 'shared-utils/assets/logos/akash-dark.svg?url';
 import assetmantleLogoDark from 'shared-utils/assets/logos/assetmantle-dark.svg?url';
@@ -88,6 +89,8 @@ import gitopiaLogoDark from 'shared-utils/assets/logos/gitopia-dark.svg?url';
 import gitopiaLogoLight from 'shared-utils/assets/logos/gitopia-light.svg?url';
 import neutronLogoDark from 'shared-utils/assets/logos/neutron-dark.svg?url';
 import neutronLogoLight from 'shared-utils/assets/logos/neutron-light.svg?url';
+import jackalLogoDark from 'shared-utils/assets/logos/jackal-dark.svg?url';
+import jackalLogoLight from 'shared-utils/assets/logos/jackal-light.svg?url';
 
 interface IconProps extends Omit<ImageProps, 'id' | 'src'> {
   type: 'icon' | 'logo';
@@ -281,6 +284,10 @@ const ChainIcon = ({
     case 'neutron':
       [iconDark, iconLight] =
         type === 'icon' ? [neutronIconBoth, neutronIconBoth] : [neutronLogoDark, neutronLogoLight];
+      break;
+    case 'jackal':
+      [iconDark, iconLight] =
+        type === 'icon' ? [jackalIconBoth, jackalIconBoth] : [jackalLogoDark, jackalLogoLight];
       break;
     default:
       throw new Error(`chain ${chainName} not supported`);